refactor : Favorite과 Member 일대다 단방향 매핑
